refactor(services): split evaluations index, mixin, mutation and accumulated
Task 14 (Phase 2 finale) of the size-limits burndown.

_evaluations_index.py: status.json readers (dimensions/deadline/provider-model/
time-limit/tail-log) plus the RunRow -> JobSnapshot assembly move to
_run_status_readers.py; list()'s internal-jobs merge extracts to
_merge_internal_jobs. Facade stays at 373L (over the 300 cap): the brief
scoped a single extraction target for a 359-line class, matching the Task 13
"class stays whole" precedent for jobs.py under the extraction-overhead rule.

evaluation_mixin.py splits into _evaluation_dispatch.py (EvaluationDispatcher,
SubprocessDispatcher, _build_evaluate_cmd), _evaluation_env.py (free
build_eval_env, mixin delegates), and _run_discard.py (cancel-wait + discard
machinery). All re-exported; get_ai_cmd/get_ai_model resolution stays in the
facade so existing patches keep working.

mutation_rescore.py splits into _mutation_projection.py (ProjectLockRegistry,
project-wide projection sweep) and _mutation_scoring.py (slim rescore payload,
default-run resolution). ProjectLockRegistry/_DEFAULT_PROJECT_LOCKS and
_project_all_runs re-exported for direct test imports/patches.

accumulated.py splits into _accumulated_cache.py (walk-cache globals, per-call
LRU config) and _accumulated_aggregate.py (trend/severity/score aggregation,
including the to_camel_dict wire-serialization call). DECLARED_WIRE_BOUNDARIES
updated to point at the new call site.

Added scope (Task 13 leftovers): _fs_metadata._read_accumulated_summary and
_fs_project_helpers._build_project_entry decomposed under 50 lines each via
intra-file helper extraction.

Baseline: 158 -> 152 grandfathered entries.
