+++ tools/size_baseline.txt
@@ -91,18 +91,12 @@ src/quodeq/data/sqlite/run_index.py:93:function
 src/quodeq/data/sqlite/state_store.py:1:file
 src/quodeq/services/_cache.py:149:function
 src/quodeq/services/_evaluations_index.py:1:file
-src/quodeq/services/_evaluations_index.py:59:function
-src/quodeq/services/_fs_metadata.py:160:function
-src/quodeq/services/_fs_project_helpers.py:74:function
 src/quodeq/services/_job_model.py:1:file
 src/quodeq/services/_trend_fetcher.py:82:function
-src/quodeq/services/accumulated.py:1:file
 src/quodeq/services/dashboard.py:183:function
 src/quodeq/services/dismissed.py:121:function
-src/quodeq/services/evaluation_mixin.py:1:file
 src/quodeq/services/evidence_rescore.py:46:function
 src/quodeq/services/jobs.py:1:file
-src/quodeq/services/mutation_rescore.py:1:file
 src/quodeq/services/scan_progress.py:1:file
 src/quodeq/services/scan_progress.py:257:function
 src/quodeq/services/score_run.py:50:function
